requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
qrcode>=7.4.0
pyotp>=2.8.0
psutil>=5.9.0
pyperclip>=1.8.0
# Dependencia para System Tray (Segundo plano)
pystray>=0.19.0
# Dependencias específicas de Windows
pywin32>=306
pypiwin32>=223
# Para compilar
pyinstaller>=6.0.0
# Para el sistema de actualizaciones
requests>=2.31.0
packaging>=23.0
# Para dar formato al código
black>=23.0.0